del mini-side-nav; copy to new.js
