+++ src/utils/constants.js
@@ -17,3 +17,4 @@ export const PAGE_NOT_FOUND = "*";
 export const ALGORITHMS = "algorithms";
 export const SCRIPTING = "scripting";
 export const SQL = "sql";
+export const SAVE = "save";
